+++ weaver/wps_restapi/__init__.py
@@ -4,8 +4,7 @@ from typing import TYPE_CHECKING
 from pyramid.settings import asbool
 
 from weaver.utils import get_settings
-from weaver.wps_restapi.swagger_definitions import cornice_route_prefix
-from weaver.wps_restapi.utils import wps_restapi_base_path
+from weaver.wps_restapi.utils import get_wps_restapi_base_path
 
 if TYPE_CHECKING:
     from pyramid.config import Configurator
@@ -22,8 +21,8 @@ def includeme(config):
         LOGGER.warning("Skipping WPS REST API views [weaver.wps_restapi=false].")
     else:
         LOGGER.info("Adding WPS REST API views...")
-        api_base = wps_restapi_base_path(settings)
-        with cornice_route_prefix(config, route_prefix=api_base):
+        api_base = get_wps_restapi_base_path(settings)
+        with config.route_prefix_context(route_prefix=api_base):
             config.include("weaver.wps_restapi.jobs")
             config.include("weaver.wps_restapi.providers")
             config.include("weaver.wps_restapi.processes")